+++ azext_iot/common/utility.py
@@ -438,10 +438,14 @@ class ISO8601Validator:
             return False
 
 
-def ensure_min_version(cur_ver, min_ver):
-    from pkg_resources._vendor.packaging import version
+def ensure_iothub_sdk_min_version(min_ver):
+    from packaging import version
+    try:
+        from azure.mgmt.iothub import __version__ as iot_sdk_version
+    except ImportError:
+        from azure.mgmt.iothub._configuration import VERSION as iot_sdk_version
 
-    return version.parse(cur_ver) >= version.parse(min_ver)
+    return version.parse(iot_sdk_version) >= version.parse(min_ver)
 
 
 def scantree(path):