+++ azext_iot/digitaltwins/providers/twin.py
@@ -89,8 +89,10 @@ class TwinProvider(DigitalTwinsProvider):
         json_patch_collection = []
         if isinstance(json_patch, dict):
             json_patch_collection.append(json_patch)
-        if isinstance(json_patch, list):
+        elif isinstance(json_patch, list):
             json_patch_collection.extend(json_patch)
+        else:
+            raise CLIError(f"--json-patch content must be an object or array. Actual type was: {type(json_patch).__name__}")
 
         logger.info("Patch payload %s", json.dumps(json_patch_collection))
 
@@ -104,12 +106,28 @@ class TwinProvider(DigitalTwinsProvider):
             raise CLIError(unpack_msrest_error(e))
 
     def delete(self, twin_id, etag=None):
-        # Not a json response
         try:
             options = TwinOptions(if_match=(etag if etag else "*"))
             self.twins_sdk.delete(id=twin_id, digital_twins_delete_options=options, raw=True)
         except ErrorResponseException as e:
             raise CLIError(unpack_msrest_error(e))
+
+    def delete_all(self, only_relationships=False):
+        # need to get all twins
+        query = "select * from digitaltwins"
+        twins = self.invoke_query(query=query, show_cost=False)["result"]
+        print(f"Found {len(twins)} twin(s).")
+
+        # go through and delete all
+        for twin in twins:
+            try:
+                self.delete_all_relationship(
+                    twin_id=twin["$dtId"]
+                )
+                if not only_relationships:
+                    self.delete(twin_id=twin["$dtId"])
+            except CLIError as e:
+                logger.warn(f"Could not delete twin {twin['$dtId']}. The error is {e}")
 
     def add_relationship(
         self,
@@ -186,8 +204,10 @@ class TwinProvider(DigitalTwinsProvider):
         json_patch_collection = []
         if isinstance(json_patch, dict):
             json_patch_collection.append(json_patch)
-        if isinstance(json_patch, list):
+        elif isinstance(json_patch, list):
             json_patch_collection.extend(json_patch)
+        else:
+            raise CLIError(f"--json-patch content must be an object or array. Actual type was: {type(json_patch).__name__}")
 
         logger.info("Patch payload %s", json.dumps(json_patch_collection))
 
@@ -214,6 +234,34 @@ class TwinProvider(DigitalTwinsProvider):
         except ErrorResponseException as e:
             raise CLIError(unpack_msrest_error(e))
 
+    def delete_all_relationship(self, twin_id):
+        relationships = self.list_relationships(twin_id, incoming_relationships=True)
+        incoming_pager = self.list_relationships(twin_id)
+
+        # relationships pager needs to be advanced to get relationships
+        try:
+            while True:
+                relationships.extend(incoming_pager.advance_page())
+        except StopIteration:
+            pass
+
+        print(f"Found {len(relationships)} relationship(s) associated with twin {twin_id}.")
+
+        for relationship in relationships:
+            try:
+                if isinstance(relationship, dict):
+                    self.delete_relationship(
+                        twin_id=twin_id,
+                        relationship_id=relationship['$relationshipId']
+                    )
+                else:
+                    self.delete_relationship(
+                        twin_id=relationship.source_id,
+                        relationship_id=relationship.relationship_id
+                    )
+            except CLIError as e:
+                logger.warn(f"Could not delete relationship {relationship}. The error is {e}.")
+
     def get_component(self, twin_id, component_path):
         try:
             return self.twins_sdk.get_component(
@@ -228,8 +276,10 @@ class TwinProvider(DigitalTwinsProvider):
         json_patch_collection = []
         if isinstance(json_patch, dict):
             json_patch_collection.append(json_patch)
-        if isinstance(json_patch, list):
+        elif isinstance(json_patch, list):
             json_patch_collection.extend(json_patch)
+        else:
+            raise CLIError(f"--json-patch content must be an object or array. Actual type was: {type(json_patch).__name__}")
 
         logger.info("Patch payload %s", json.dumps(json_patch_collection))