+++ azext_iot/iothub/providers/base.py
@@ -21,7 +21,7 @@ class IoTHubProvider(object):
         self.discovery = IotHubDiscovery(cmd)
         self.target = self.discovery.get_target(
             hub_name=self.hub_name,
-            rg=self.rg,
+            resource_group_name=self.rg,
             login=login,
         )
         self.resolver = SdkResolver(self.target)